# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api authentification",
        "description": "Implementation authentification api",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/signin": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion",
                "operationId": "66bf0f7f83e7ed324b1b6e1fab9cb49d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès"
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Création de compte utilisateur",
                "operationId": "47fd319b407ed1fae5f24abf9ed83302",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lastname",
                                    "firstname",
                                    "username",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone_number",
                                    "type"
                                ],
                                "properties": {
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès"
                    }
                }
            }
        },
        "/api/signout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion sécurisée de l'utilisateur",
                "operationId": "59c454b60c0676ec0dcea2b2e4761a38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Changement de mot de passe",
                "operationId": "bbf02faac4cfbb998aad2b63a843b88c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "old_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "old_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentification",
            "description": "Authentification"
        }
    ]
}